extracted metadata related macro code to separate trait
